stock bar update
