+++ src/App.js
@@ -1,7 +1,6 @@
 import './sass/App.css';
 import Header from './components/Header/header';
 import Stockbar from './components/StockBar/stockbar';
-import Stockbar2 from './components/StockBar2/stockbar2';
 import Main from './components/Main/main';
 import Footer from './components/Footer/footer';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
@@ -11,7 +10,7 @@ function App() {
     <Router>
         <div className="App">
           <Header />
-          <Stockbar2 />
+          <Stockbar />
           <Main />            
           <Footer />
         </div>
