+++ src/components/StockBarAlt/css/customCSS.css
