+++ src/components/StockBarAlt/sass/stockbarAlt.css
@@ -180,4 +180,4 @@
 .outerMarquee .companyRow {
   border-bottom: none !important;
 }
-/*# sourceMappingURL=stockbar2.css.map */
+/*# sourceMappingURL=stockbarAlt.css.map */
