+++ src/components/StockBarAlt/sass/stockbarAlt.scss
